Add SyntaxQ and add String Tests section...
of a new "Testing Expressions" Guide Section
