+++ .pre-commit-config.yaml
@@ -6,17 +6,17 @@ repos:
     hooks:
     - id: check-merge-conflict
     - id: debug-statements
-      stages: [commit]
+      stages: [pre-commit]
     - id: end-of-file-fixer
-      stages: [commit]
+      stages: [pre-commit]
 -   repo: https://github.com/pycqa/isort
     rev: 5.13.2
     hooks:
       - id: isort
-        stages: [commit]
+        stages: [pre-commit]
 -   repo: https://github.com/psf/black
     rev: 23.12.1
     hooks:
     - id: black
       language_version: python3
-      stages: [commit]
+      stages: [pre-commit]
