+++ mathics/builtin/string/__init__.py
@@ -2,3 +2,4 @@
 Strings and Characters
 
 """
+# FIXME: Redo. This is a Tech note, not a Guide Section.
