+++ mathics/core/systemsymbols.py
@@ -257,6 +257,7 @@ SymbolStringExpression = Symbol("System`StringExpression")
 SymbolStringForm = Symbol("System`StringForm")
 SymbolStringInsert = Symbol("System`StringInsert")
 SymbolStringJoin = Symbol("System`StringJoin")
+SymbolStringMatchQ = Symbol("System`StringMatchQ")
 SymbolStringPosition = Symbol("System`StringPosition")
 SymbolStringQ = Symbol("System`StringQ")
 SymbolStringRiffle = Symbol("System`StringRiffle")
@@ -276,6 +277,7 @@ SymbolTeXForm = Symbol("System`TeXForm")
 SymbolTimes = Symbol("System`Times")
 SymbolThrow = Symbol("System`Throw")
 SymbolThreshold = Symbol("System`Threshold")
+SymbolToExpression = Symbol("System`ToExpression")
 SymbolToString = Symbol("System`ToString")
 SymbolTotal = Symbol("System`Total")
 SymbolTraditionalForm = Symbol("System`TraditionalForm")
